2023 day 6 part 2 win the real race by using long instead of int

+++ src/main/java/tk/vivas/adventofcode/year2023/day06/RaceRecord.java
@@ -2,9 +2,9 @@ package tk.vivas.adventofcode.year2023.day06;
 
 import static java.lang.Math.*;
 
-record RaceRecord(int duration, int distance) {
+record RaceRecord(long duration, long distance) {
 
-    int marginOfError() {
+    long marginOfError() {
         // distance < (timeToLoad) * (duration - timeToLoad)
         // distance < timeToLoad * duration - timeToLoad²
         // -d > x² - tx
@@ -18,13 +18,13 @@ record RaceRecord(int duration, int distance) {
         double x2 = (-b - sqrt) / 2;
 
         double min = min(x1, x2);
-        int minInt = (int) Math.ceil(min);
+        long minInt = (long) Math.ceil(min);
         if (minInt == min) {
             minInt++;
         }
 
         double max = max(x1, x2);
-        int maxInt = (int) Math.floor(max);
+        long maxInt = (long) Math.floor(max);
         if (maxInt == max) {
             maxInt--;
         }
